pyproject.toml
[project]
name = "stone-ocr"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (==2.1.2)",
    "torchvision (==0.16.2)",
    "torchaudio (==2.1.2)",
    "lightning (>=2.5.1,<3.0.0)",
    "segmentation-models-pytorch (>=0.5.0,<0.6.0)",
    "albumentations[opencv] (>=2.0.5,<3.0.0)",
    "torchmetrics (>=1.7.1,<2.0.0)",
    "ptflops (>=0.7.4,<0.8.0)",
    "numpy (<2)",
    "pillow (>=11.2.1,<12.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "monai (>=1.4.0,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)"
]

[tool.poetry]
packages = [{include = "stone_ocr", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
